Adds delete API and refactors package structure
