+++ sample-resource-plugin/src/main/java/org/opensearch/sample/actions/resource/delete/DeleteResourceRequest.java
@@ -1,0 +1,49 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.sample.actions.resource.delete;
+
+import java.io.IOException;
+
+import org.opensearch.action.ActionRequest;
+import org.opensearch.action.ActionRequestValidationException;
+import org.opensearch.core.common.io.stream.StreamInput;
+import org.opensearch.core.common.io.stream.StreamOutput;
+
+/**
+ * Request object for CreateSampleResource transport action
+ */
+public class DeleteResourceRequest extends ActionRequest {
+
+    private final String resourceId;
+
+    /**
+     * Default constructor
+     */
+    public DeleteResourceRequest(String resourceId) {
+        this.resourceId = resourceId;
+    }
+
+    public DeleteResourceRequest(StreamInput in) throws IOException {
+        this.resourceId = in.readString();
+    }
+
+    @Override
+    public void writeTo(final StreamOutput out) throws IOException {
+        out.writeString(this.resourceId);
+    }
+
+    @Override
+    public ActionRequestValidationException validate() {
+        return null;
+    }
+
+    public String getResourceId() {
+        return this.resourceId;
+    }
+}
